refactor: UserAccount 클래스 data.model 패키지로 이동
