+++ app/src/main/java/com/moroom/android/SignupActivity.java
@@ -11,10 +11,9 @@ import android.text.Editable;
 import android.text.TextWatcher;
 import android.util.Log;
 import android.view.View;
-import android.widget.Button;
-import android.widget.EditText;
 import android.widget.Toast;
 
+import com.moroom.android.data.model.UserAccount;
 import com.moroom.android.databinding.ActivitySignUpBinding;
 import com.google.firebase.auth.ActionCodeSettings;
 import com.google.firebase.auth.FirebaseAuth;
