+++ app/src/main/java/com/moroom/android/data/model/UserAccount.java
@@ -1,4 +1,4 @@
-package com.moroom.android;
+package com.moroom.android.data.model;
 
 public class UserAccount {
     private String idToken;
